Add basic travis-CI Support
Add basic travis-CI support. All this does is run middleman build in
lieu of a test suite. If the build fails then we have an error somewhere.

The next step is to get Travis to deploy the generated site to the beta
or main site depending on the branch name